requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
idna==3.11
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
nltk==3.9.2
numpy==2.0.2
oauthlib==3.3.1
osometweet==0.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sklearn==0.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
twitterbloc==1.2.1
tzdata==2025.2
urllib3==1.26.20
zipp==3.23.0
